style: 네트워크 파일 정리
