+++ FlagBoard/FlagBoard/Network/Manager/AlamofireManager.swift
